refactor(dataset): split _reconstruct.py + extract _query.py via QueryView (PR6)
Big move. Splits the heavyweight modules along architectural seams.

Reconstructor splits:

- New _dataset/_protocol.py (34 lines): houses the Reconstructor Protocol.
  Pulled into its own module to break the import cycle that arises when leaf
  modules each import the Protocol and the dispatcher module (_reconstruct.py)
  imports the leaves.
- New _dataset/_ref.py (76 lines): the Ref reconstructor.
- New _dataset/_haps.py (870 lines): the Haps reconstructor along with its
  supporting value types (_Variants, ReconstructionRequest).
- _dataset/_tracks.py (379 -> 736 lines): the Tracks class and TrackType enum
  merged in alongside the existing numba kernels. Class lives next to the
  helpers it dispatches to.
- _dataset/_reconstruct.py (1539 -> 310 lines): slimmed to the dispatcher
  pieces only -- RefTracks, HapsTracks, _build_reconstructor -- plus
  re-exports so callers that imported moved symbols from this module keep
  working (_impl.py, _open.py, _dummy.py).

QueryView composition:

- New _dataset/_query.py (403 lines): houses a typed QueryView frozen
  dataclass over the 9 Dataset fields that the query path needs (idxer,
  sp_idxer, full_regions, rng, recon, output_length, jitter, deterministic,
  rc_neg). The query logic moves to free functions (getitem,
  _getitem_unspliced, _getitem_spliced, build_recon_splice_plan,
  reverse_complement_ragged [was Dataset._rc], pad [was Dataset._pad],
  _regroup) that take QueryView as their first arg. Each stage is testable
  in isolation against a synthetic view.
- Dataset.__getitem__ becomes a ~15-line facade that builds a QueryView and
  delegates to _query.getitem. The "inner_ds = self.with_len('ragged')" trick
  in _getitem_spliced is replaced by passing output_length="ragged" directly
  to the recon call -- no inner Dataset construction needed.
- _impl.py: 2015 -> 1712 lines (-303).

Other:

- Reconstructor.__call__ now declares splice_plan: SplicePlan | None = None,
  matching what all implementers already accept. Previously the Protocol
  underspecified its contract; pyrefly caught it during the QueryView
  rewrite.
- Campaign spec updated inline to record audit findings before
  implementation (existing _tracks.py was numba-only, not a class home; the
  query extraction is not mechanical and benefits from composition) and
  the as-shipped design.

Behavior preserving; kernel calls unchanged; full test suite (488 pytest +
cargo) green; pyrefly + ruff clean.
