+++ python/genvarloader/_dataset/_impl.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from collections.abc import Iterable, Mapping, Sequence
 from pathlib import Path
-from typing import Callable, Generic, Literal, TypeVar, cast, overload
+from typing import Callable, Generic, Literal, TypeVar, overload
 
 import awkward as ak
 import numpy as np
@@ -11,7 +11,7 @@ import seqpro as sp
 from dataclasses import dataclass, replace
 from loguru import logger
 from numpy.typing import NDArray
-from seqpro.rag import DTYPE_co as DTYPE, Ragged, is_rag_dtype
+from seqpro.rag import Ragged
 from typing_extensions import NoReturn, Self, assert_never
 
 from .._ragged import (
@@ -20,8 +20,6 @@ from .._ragged import (
     RaggedIntervals,
     RaggedSeqs,
     RaggedTracks,
-    reverse_complement,
-    to_padded,
 )
 from .._torch import TORCH_AVAILABLE, TorchDataset, get_dataloader
 from .._types import AnnotatedHaps, Idx, StrIdx
@@ -30,7 +28,7 @@ from .._utils import (
     normalize_contig_name,
 )
 from ._indexing import DatasetIndexer, SpliceIndexer, is_str_arr
-from ._splice import SpliceMap, SplicePlan, build_splice_plan
+from ._splice import SpliceMap
 from ._rag_variants import RaggedVariants
 from ._insertion_fill import InsertionFill
 from ._reconstruct import (
@@ -1355,323 +1353,22 @@ class Dataset:
             ...,
         ]
     ):
-        if is_str_arr(idx) and self._idxer.r2i_map is None:
-            raise ValueError(
-                "Cannot query regions by name because no region name was set."
-            )
+        # Thin facade: package state into a QueryView and hand off to the
+        # query module's free functions, which carry the actual logic.
+        from ._query import QueryView, getitem
 
-        if self._sp_idxer is not None:
-            recon, squeeze, out_reshape = self._getitem_spliced(idx, self._sp_idxer)
-        else:
-            recon, squeeze, out_reshape = self._getitem_unspliced(idx)
-
-        if self.output_length == "variable":
-            recon = tuple(
-                r if isinstance(r, (RaggedVariants, RaggedIntervals)) else self._pad(r)
-                for r in recon
-            )
-        elif isinstance(self.output_length, int):
-            recon = tuple(
-                r if isinstance(r, (RaggedVariants, RaggedIntervals)) else r.to_numpy()
-                for r in recon
-            )
-
-        if out_reshape is not None:
-            recon = tuple(o.reshape(out_reshape + o.shape[1:]) for o in recon)  # type: ignore
-
-        if squeeze:
-            # (1 [p] l) -> ([p] l)
-            recon = tuple(o.squeeze(0) for o in recon)  # type: ignore
-
-        if len(recon) == 1:
-            recon = recon[0]
-
-        return recon
-
-    def _getitem_unspliced(
-        self, idx: StrIdx | tuple[StrIdx] | tuple[StrIdx, StrIdx]
-    ) -> tuple[
-        tuple[
-            Ragged[np.bytes_ | np.float32] | RaggedAnnotatedHaps | RaggedVariants, ...
-        ],
-        bool,
-        tuple[int, ...] | None,
-    ]:
-        # (b)
-        ds_idx, squeeze, out_reshape = self._idxer.parse_idx(idx)
-        r_idx, _ = np.unravel_index(ds_idx, self.full_shape)
-
-        # makes a copy because r_idx is at least 1D & triggers advanced indexing
-        regions = self._full_regions[r_idx]
-        lengths = regions[:, 2] - regions[:, 1]
-        jitter = self._rng.integers(
-            -self.jitter, self.jitter + 1, size=len(regions), dtype=np.int32
-        )
-        regions[:, 1] += jitter
-        regions[:, 2] = regions[:, 1] + lengths
-
-        recon = self._recon(
-            idx=ds_idx,
-            r_idx=r_idx,
-            regions=regions,
+        view = QueryView(
+            idxer=self._idxer,
+            sp_idxer=self._sp_idxer,
+            full_regions=self._full_regions,
+            rng=self._rng,
+            recon=self._recon,
             output_length=self.output_length,
             jitter=self.jitter,
-            rng=self._rng,
             deterministic=self.deterministic,
+            rc_neg=self.rc_neg,
         )
-
-        if not isinstance(recon, tuple):
-            recon = (recon,)
-
-        if self.rc_neg:
-            to_rc: NDArray[np.bool_] = self._full_regions[r_idx, 3] == -1
-            recon = tuple(self._rc(r, to_rc) for r in recon)
-
-        return recon, squeeze, out_reshape  # type: ignore
-
-    def _getitem_spliced(
-        self,
-        idx: StrIdx | tuple[StrIdx] | tuple[StrIdx, StrIdx],
-        splice_idxer: SpliceIndexer,
-    ) -> tuple[
-        tuple[Ragged[np.bytes_ | np.float32] | RaggedAnnotatedHaps, ...],
-        bool,
-        tuple[int, ...] | None,
-    ]:
-        if isinstance(self.output_length, int):
-            raise RuntimeError(
-                "In general, splicing cannot be done with fixed length data because even if the length of each region's data"
-                " is fixed/constant, the number of elements in each spliced element is not. Thus, the final length of the"
-                " spliced elements will be variable."
-            )
-
-        assert self.sequence_type != "variants"
-        assert not isinstance(self.output_length, int)
-        assert self.jitter == 0
-        assert self.deterministic
-
-        # TODO: really need to assert no jitter and deterministic?
-        # * In theory, this still "works" with jitter or non-determinism, but why would anyone want this? Would they want a different alg here?
-        # * Potential issues:
-        # * Each each component of the spliced output will have different jitter
-        # * For non-determinism, each component will have different shifts & different unphased haplotypes chosen
-        if self.jitter > 0:
-            raise RuntimeError(
-                "Jitter is not supported with splicing. Please set jitter to 0."
-            )
-
-        if not self.deterministic:
-            raise RuntimeError(
-                "Non-deterministic algorithms are not supported with splicing. Please set deterministic to True."
-            )
-
-        inner_ds = self.with_len("ragged")
-        (
-            ds_idx,
-            squeeze,
-            out_reshape,
-            offsets,
-            n_rows_sel,
-            n_samples_sel,
-        ) = splice_idxer.parse_idx(idx)
-        r_idx, _ = np.unravel_index(ds_idx, self._idxer.full_shape)
-        regions = self._full_regions[r_idx]
-
-        # Build the splice plan from per-query lengths produced by the active
-        # reconstructor. The plan drives the kernel into writing pre-spliced
-        # bytes in (splice_row, sample, *inner_fixed, splice_element) C-order.
-        plan = inner_ds._build_splice_plan(
-            ds_idx=ds_idx,
-            regions=regions,
-            splice_row_offsets=offsets,
-            n_rows=n_rows_sel,
-            n_samples=n_samples_sel,
-        )
-
-        recon = inner_ds._recon(
-            idx=ds_idx,
-            r_idx=r_idx,
-            regions=regions,
-            output_length="ragged",
-            jitter=self.jitter,
-            rng=self._rng,
-            deterministic=self.deterministic,
-            splice_plan=plan,
-        )
-
-        if not isinstance(recon, tuple):
-            recon = (recon,)
-
-        recon = cast(
-            tuple[Ragged[np.bytes_ | np.float32] | RaggedAnnotatedHaps, ...], recon
-        )
-
-        if self.rc_neg:
-            # Permute the per-region to_rc mask the same way the plan permuted
-            # the kernel queries. The plan acts on a flattened (B, *inner_fixed)
-            # k-index, so first replicate to_rc across the inner axes, then
-            # gather via plan.perm.
-            B = regions.shape[0]
-            n_k = int(plan.perm.shape[0])
-            inner_factor, rem = divmod(n_k, B)
-            if rem != 0:
-                raise AssertionError(
-                    "plan.perm length is not a multiple of len(regions); "
-                    "inner-fixed flatten factor inconsistent."
-                )
-            to_rc_unperm = regions[:, 3] == -1
-            if inner_factor == 1:
-                to_rc_flat = to_rc_unperm
-            else:
-                # (B, E) C-order: same value across the inner axis for a given
-                # query. np.repeat gives (B*E,) in (query, inner) C-order.
-                to_rc_flat = np.repeat(to_rc_unperm, inner_factor)
-            to_rc_per_elem: NDArray[np.bool_] = to_rc_flat[plan.perm]
-            recon = tuple(self._rc(r, to_rc_per_elem) for r in recon)
-
-        # Rewrap each per-element Ragged with the plan's group_offsets to
-        # expose one contiguous spliced element per (row, sample[, inner]) cell.
-        # Collapse (n_rows, n_samples) into a single leading "pair" axis so the
-        # downstream out_reshape step in __getitem__ can reshape it back to
-        # whatever shape the user's index requested.
-        recon = tuple(
-            _regroup(r, plan.group_offsets, plan.flat_out_shape) for r in recon
-        )
-
-        return recon, squeeze, out_reshape
-
-    def _build_splice_plan(
-        self,
-        ds_idx: NDArray[np.intp],
-        regions: NDArray[np.int32],
-        splice_row_offsets: NDArray[np.int64],
-        n_rows: int,
-        n_samples: int,
-    ) -> SplicePlan:
-        """Build a ``SplicePlan`` for the active reconstructor.
-
-        Dispatches on ``type(self._recon)``: ``Haps`` uses (B, P) haplotype
-        lengths; ``Ref`` uses 1-D per-region lengths. Track-bearing
-        reconstructors are not yet supported in the spliced path.
-        """
-        recon = self._recon
-        if isinstance(recon, HapsTracks):
-            raise NotImplementedError(
-                "Splicing of haplotypes + tracks (shape (b, t, p, ~l)) is not "
-                "supported."
-            )
-        if isinstance(recon, RefTracks):
-            raise NotImplementedError(
-                "Splicing of reference + tracks is not yet supported."
-            )
-        if isinstance(recon, Tracks):
-            # Tracks have deterministic per-region lengths (no haplotype
-            # indels). Replicate the (B,) length array across the n_tracks
-            # inner axis so the plan's inner_fixed = (n_tracks,).
-            n_tracks = len(recon.active_tracks)
-            base = (regions[:, 2] - regions[:, 1]).astype(np.int32, copy=False)
-            lengths_2d = np.broadcast_to(base[:, None], (base.shape[0], n_tracks))
-            return build_splice_plan(
-                lengths=np.ascontiguousarray(lengths_2d),
-                splice_row_offsets=splice_row_offsets,
-                n_samples=n_samples,
-                n_rows=n_rows,
-            )
-        if isinstance(recon, Haps):
-            lengths_2d = recon.haplotype_lengths_for_plan(idx=ds_idx, regions=regions)
-            return build_splice_plan(
-                lengths=lengths_2d.astype(np.int32, copy=False),
-                splice_row_offsets=splice_row_offsets,
-                n_samples=n_samples,
-                n_rows=n_rows,
-            )
-        if isinstance(recon, Ref):
-            lengths_1d = (regions[:, 2] - regions[:, 1]).astype(np.int32, copy=False)
-            return build_splice_plan(
-                lengths=lengths_1d,
-                splice_row_offsets=splice_row_offsets,
-                n_samples=n_samples,
-                n_rows=n_rows,
-            )
-        raise NotImplementedError(
-            f"Splicing not supported for reconstructor {type(recon).__name__}."
-        )
-
-    @overload
-    def _rc(self, rag: Ragged[DTYPE], to_rc: NDArray[np.bool_]) -> Ragged[DTYPE]: ...
-    @overload
-    def _rc(
-        self, rag: RaggedAnnotatedHaps, to_rc: NDArray[np.bool_]
-    ) -> RaggedAnnotatedHaps: ...
-    @overload
-    def _rc(self, rag: RaggedVariants, to_rc: NDArray[np.bool_]) -> RaggedVariants: ...
-    @overload
-    def _rc(
-        self, rag: RaggedIntervals, to_rc: NDArray[np.bool_]
-    ) -> RaggedIntervals: ...
-    def _rc(
-        self,
-        rag: Ragged | RaggedAnnotatedHaps | RaggedVariants | RaggedIntervals,
-        to_rc: NDArray[np.bool_],
-    ) -> Ragged | RaggedAnnotatedHaps | RaggedVariants | RaggedIntervals:
-        if isinstance(rag, Ragged):
-            if is_rag_dtype(rag, np.bytes_):
-                rag = Ragged(
-                    ak.to_packed(ak.where(to_rc, reverse_complement(rag), rag))
-                )  # type: ignore
-            else:
-                rag = Ragged(ak.to_packed(ak.where(to_rc, rag[..., ::-1], rag)))  # type: ignore
-        elif isinstance(rag, RaggedAnnotatedHaps):
-            rag.haps = self._rc(rag.haps, to_rc)
-            rag.var_idxs = self._rc(rag.var_idxs, to_rc)
-            rag.ref_coords = self._rc(rag.ref_coords, to_rc)
-        elif isinstance(rag, RaggedVariants):
-            rag = rag.rc_(to_rc)
-        elif isinstance(rag, RaggedIntervals):
-            rag = rag
-        else:
-            assert_never(rag)
-        return rag
-
-    @overload
-    def _pad(self, rag: Ragged[DTYPE]) -> NDArray[DTYPE]: ...
-    @overload
-    def _pad(self, rag: RaggedAnnotatedHaps) -> AnnotatedHaps: ...
-    def _pad(self, rag: Ragged | RaggedAnnotatedHaps) -> NDArray | AnnotatedHaps:
-        if isinstance(rag, Ragged):
-            if is_rag_dtype(rag, np.bytes_):
-                return to_padded(rag, b"N")
-            elif is_rag_dtype(rag, np.float32):
-                return to_padded(rag, 0)
-            else:
-                raise ValueError(f"Unsupported pad dtype: {rag.data.dtype}")
-        elif isinstance(rag, RaggedAnnotatedHaps):
-            return rag.to_padded()
-        else:
-            assert_never(rag)
-
-
-def _regroup(
-    rag: Ragged | RaggedAnnotatedHaps,
-    group_offsets: NDArray[np.int64],
-    out_shape: tuple[int | None, ...],
-) -> Ragged | RaggedAnnotatedHaps:
-    """Rewrap a per-element Ragged (or RaggedAnnotatedHaps) with grouped
-    offsets so each cell holds one contiguous spliced element.
-
-    Both branches share the same data buffer; only the outer offsets / shape
-    change.
-    """
-    if isinstance(rag, RaggedAnnotatedHaps):
-        return RaggedAnnotatedHaps(
-            haps=cast(
-                Ragged[np.bytes_],
-                _regroup(rag.haps, group_offsets, out_shape),
-            ),
-            var_idxs=cast(Ragged, _regroup(rag.var_idxs, group_offsets, out_shape)),
-            ref_coords=cast(Ragged, _regroup(rag.ref_coords, group_offsets, out_shape)),
-        )
-    return Ragged.from_offsets(rag.data, out_shape, group_offsets)
+        return getitem(view, idx)
 
 
 def _annot_to_intervals(regions: pl.DataFrame, annot: pl.DataFrame) -> RaggedIntervals:
